+++ gnfinder_test.go
@@ -20,8 +20,9 @@ var _ = Describe("GNfinder", func() {
 	Describe("NewGNfinder()", func() {
 		It("returns new GNfinder object", func() {
 			gnf := NewGNfinder()
-			Expect(gnf.LanguageUsed).To(Equal(lang.NotSet))
-			Expect(gnf.Bayes).To(BeFalse())
+			Expect(gnf.Language).To(Equal(lang.DefaultLanguage))
+			Expect(gnf.LanguageDetected).To(Equal(""))
+			Expect(gnf.Bayes).To(BeTrue())
 			Expect(gnf.Verifier).To(BeNil())
 			// dictionary is loaded internally
 			Expect(len(gnf.Dict.Ranks)).To(BeNumerically(">", 5))
@@ -30,13 +31,14 @@ var _ = Describe("GNfinder", func() {
 		It("takes language", func() {
 			gnf := NewGNfinder(OptDict(dictionary), OptBayesWeights(weights),
 				OptLanguage(lang.English))
-			Expect(gnf.LanguageUsed).To(Equal(lang.English))
+			Expect(gnf.Language).To(Equal(lang.English))
+			Expect(gnf.DetectLanguage).To(BeFalse())
+			Expect(gnf.LanguageDetected).To(Equal(""))
 		})
 
 		It("sets bayes", func() {
-			gnf := NewGNfinder(OptDict(dictionary), OptBayesWeights(weights),
-				OptBayes(true))
-			Expect(gnf.Bayes).To(BeTrue())
+			gnf := NewGNfinder(OptBayes(false))
+			Expect(gnf.Bayes).To(BeFalse())
 		})
 
 		It("sets bayes' threshold", func() {
@@ -56,13 +58,42 @@ var _ = Describe("GNfinder", func() {
 				OptBayesWeights(weights),
 				OptVerify(vOpts...),
 				OptBayes(true),
-				OptLanguage(lang.English),
+				OptLanguage(lang.German),
 			}
 			gnf := NewGNfinder(opts...)
 			Expect(gnf.Verifier.Workers).To(Equal(10))
 			Expect(gnf.Verifier.URL).To(Equal(url))
-			Expect(gnf.LanguageUsed).To(Equal(lang.English))
+			Expect(gnf.Language).To(Equal(lang.German))
+			Expect(gnf.DetectLanguage).To(BeFalse())
 			Expect(gnf.Bayes).To(BeTrue())
+		})
+
+		Describe("Update", func() {
+			It("updates gnf returning backup", func() {
+				opts := []Option{
+					OptDict(dictionary),
+					OptBayesWeights(weights),
+					OptLanguage(lang.German),
+				}
+				gnf := NewGNfinder(opts...)
+				Expect(gnf.Language).To(Equal(lang.German))
+				Expect(gnf.DetectLanguage).To(BeFalse())
+				Expect(gnf.Bayes).To(BeTrue())
+				opts2 := []Option{
+					OptDetectLanguage(true),
+					OptBayes(false),
+				}
+				backup := gnf.Update(opts2...)
+				Expect(gnf.Language).To(Equal(lang.NotSet))
+				Expect(gnf.DetectLanguage).To(BeTrue())
+				Expect(gnf.Bayes).To(BeFalse())
+				for _, opt := range backup {
+					opt(gnf)
+				}
+				Expect(gnf.Language).To(Equal(lang.German))
+				Expect(gnf.DetectLanguage).To(BeFalse())
+				Expect(gnf.Bayes).To(BeTrue())
+			})
 		})
 	})
 })
@@ -72,29 +103,22 @@ var _ = Describe("GNfinder", func() {
 // do not use -run=XXX or -run=^$, we need tests to preload dictionary and
 // Bayes weights.
 
-// BenchmarkSmallNoBayesText runs only heuristic algorithms on small text
+// BenchmarkSmallNoBayes runs only heuristic algorithm on small text
 // without language detection
-func BenchmarkSmallNoBayesText(b *testing.B) {
-	l, err := lang.NewLanguage("eng")
-	if err != nil {
-		panic(err)
-	}
+func BenchmarkSmallNoBayes(b *testing.B) {
 	opts := []Option{
 		OptBayes(false),
-		OptLanguage(l),
 		OptDict(dictionary),
-		OptBayesWeights(weights),
 	}
 	traceFile := "small.trace"
 	input := []byte("Pardosa moesta")
 	runBenchmark(b, input, traceFile, opts)
 }
 
-// BenchmarkSmallYesBayesText runs only both algorithms on small text
-// WITH language detection
-func BenchmarkSmallYesBayesText(b *testing.B) {
+// BenchmarkSmallYesBayes runs both algorithms on small text
+// without language detection
+func BenchmarkSmallYesBayes(b *testing.B) {
 	opts := []Option{
-		OptBayes(true),
 		OptDict(dictionary),
 		OptBayesWeights(weights),
 	}
@@ -103,33 +127,38 @@ func BenchmarkSmallYesBayesText(b *testing.B) {
 	runBenchmark(b, input, traceFile, opts)
 }
 
-// BenchmarkSmallEngText runs both algorithms on small text
-// WITHOUT language detection
-func BenchmarkSmallEngText(b *testing.B) {
-	l, err := lang.NewLanguage("eng")
-	if err != nil {
-		panic(err)
-	}
+// BenchmarkSmallYesBayesLangDetect runs both algorithms on small text
+// with language detection
+func BenchmarkSmallYesBayesLangDetect(b *testing.B) {
 	opts := []Option{
-		OptLanguage(l),
 		OptDict(dictionary),
 		OptBayesWeights(weights),
+		OptDetectLanguage(true),
 	}
 	traceFile := "small-eng.trace"
 	input := []byte("Pardosa moesta")
 	runBenchmark(b, input, traceFile, opts)
 }
 
-// BenchmarkBigNoBayesText runs only heuristic algorithm on large text
-// WITHOUT language detection
-func BenchmarkBigNoBayesText(b *testing.B) {
-	l, err := lang.NewLanguage("eng")
-	if err != nil {
-		panic(err)
-	}
+// BenchmarkBigNoBayes runs only heuristic algorithm on large text
+// without language detection
+func BenchmarkBigNoBayes(b *testing.B) {
 	opts := []Option{
 		OptBayes(false),
-		OptLanguage(l),
+		OptDict(dictionary),
+	}
+	traceFile := "big.trace"
+	input, err := ioutil.ReadFile("testdata/seashells_book.txt")
+	if err != nil {
+		panic(err)
+	}
+	runBenchmark(b, input, traceFile, opts)
+}
+
+// BenchmarkBigYesBayes runs both algorithms on large text
+// without language detection
+func BenchmarkBigYesBayes(b *testing.B) {
+	opts := []Option{
 		OptDict(dictionary),
 		OptBayesWeights(weights),
 	}
@@ -141,31 +170,11 @@ func BenchmarkBigNoBayesText(b *testing.B) {
 	runBenchmark(b, input, traceFile, opts)
 }
 
-// BenchmarkBigYesBayesText runs both algorithms on large text
-// WITH language detection
-func BenchmarkBigYesBayesText(b *testing.B) {
+// BenchmarkBigYesBayesLangDetect runs both algorithms on large text
+// with language detection
+func BenchmarkBigYesBayesLangDetect(b *testing.B) {
 	opts := []Option{
-		OptBayes(true),
-		OptDict(dictionary),
-		OptBayesWeights(weights),
-	}
-	traceFile := "big.trace"
-	input, err := ioutil.ReadFile("testdata/seashells_book.txt")
-	if err != nil {
-		panic(err)
-	}
-	runBenchmark(b, input, traceFile, opts)
-}
-
-// BenchmarkBigEngText runs both algorithms on large text
-// WITHOUT language detection
-func BenchmarkBigEngText(b *testing.B) {
-	l, err := lang.NewLanguage("eng")
-	if err != nil {
-		panic(err)
-	}
-	opts := []Option{
-		OptLanguage(l),
+		OptDetectLanguage(true),
 		OptDict(dictionary),
 		OptBayesWeights(weights),
 	}